pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stadstream"
version = "0.1.0"
description = "Unified sensor temporal action detection baselines for WiFiTAD and streaming TAL experiments."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "numpy",
  "pandas",
  "pyyaml",
  "torch",
  "tqdm",
]

[tool.setuptools.packages.find]
where = ["src"]

